Create README.md
This project aims to predict the prices of cars based on features such as brand, model, year, title status, color, state, and condition. The dataset used for this project is the "USA Cars Dataset". The project involves data preprocessing, feature engineering, and building a regression model to make accurate predictions. Various machine learning techniques are explored and evaluated.